:recycle: Refactor booster packs imports with __init__.py
